requirements.txt
asgiref==3.5.2
Django==4.1.2
django-autoslug==1.9.8
django-render-partial==0.4
Pillow==9.2.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
tzdata==2022.4

six~=1.16.0
djangorestframework~=3.14.0